# perf: serialize bytes